+++ CHANGELOG.md
@@ -23,11 +23,14 @@
   - [#3687](https://github.com/wix/ricos/pull/3687) fix polls schema error
 - `unsupported-blocks`
   - [#3688](https://github.com/wix/ricos/pull/3688) default label instead of prop
+- `adsense`
+  - [#3690](https://github.com/wix/ricos/pull/3690) fix adsense doesnt show in the editor
 
 </details>
   <hr/>
 
 ## 8.70.31 (Mar 31, 2022)
+
 ### :bug: Bug Fix
 
 - `ricos-content`